delete cart double click modified

+++ controllers/userController.ts
@@ -179,14 +179,14 @@ const updateCart = catchAsync(
                 itemID = item.id;
             }
         });
-        if (contains || amount == 0) {
+        if (contains && amount == 0) {
             await prisma.cart.delete({
                 where: {
                     id: itemID,
                     productId: productId,
                 },
             });
-        } else {
+        } else if (!contains) {
             await prisma.cart.create({
                 data: {
                     ...cartInfo,
